Version 1.0.7
- Added retries for initial serial get (see below)
- Added `fatal` mode which emits errors to STDERR and produces a non-zero exit code.
- Removed extra line breaks in error logs.

check serial on S3:
additional retries here (beyond those offered in pixl-server-storage) due to AWS
metadata-based auth subsystem (e.g. 169.254.169.254) -- it can randomly fail,
and it is NOT retried as part of the AWS-SDK retry mechanism.  Since this initial
request is almost always the one that fails, we wrap this one with retries.		https://github.com/aws/aws-sdk-js-v3/issues/4407